requirements.txt
amqp==5.0.2
apiclient==1.0.4
appdirs==1.4.3
asgiref==3.3.1
attrs==20.2.0
Automat==20.2.0
Babel==2.8.0
billiard==3.6.3.0
celery==5.0.4
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-default-group==1.2.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
ConfigArgParse==1.2.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==3.2.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.3.0
django-crispy-forms==1.9.0
django-debug-toolbar==3.1.1
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-queryset-csv==1.1.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
drf-yasg==1.17.1
fcm-django==0.3.4
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
future==0.18.2
gevent==20.9.0
geventhttpclient==1.4.4
greenlet==0.4.17
httpie==2.1.0
hyperlink==20.0.1
idna==2.9
ifaddr==0.1.7
incremental==17.5.0
inflection==0.4.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
locust==1.3.2
markdown-table==2019.4.13
MarkupSafe==1.1.1
mkdir==2019.4.13
mod-wsgi==4.7.1
msgpack==1.0.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.19.2
openapi-codec==1.3.2
opencv-python==4.4.0.44
orderdict==2019.9.25
packaging==20.3
phonenumbers==8.12.2
prompt-toolkit==3.0.8
psutil==5.7.3
psycopg2==2.8.6
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
PyTurboJPEG @ git+git://github.com/lilohuang/PyTurboJPEG.git@614e1edd155f468ecce10ccfa8efba82d960d1ae
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.2
random-username==1.0.2
readme-docstring==2019.4.13
render==1.0.0
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
setupcfg==2019.4.13
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.7
task==0.2.0
treq==20.9.0
twilio==6.47.0
txFCM==0.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
values==2019.4.13
vine==5.0.0
virtualenv==20.0.18
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
write==2019.4.13
yeelight==0.5.4
zope.event==4.5.0
zope.interface==5.2.0
